FIR checker: Separate FirExternalDeclarationChecker as it is JVM-specific.
