+++ compiler/fir/entrypoint/src/org/jetbrains/kotlin/fir/checkers/CheckersContainers.kt
@@ -16,3 +16,8 @@ fun FirSessionFactory.FirSessionConfigurator.registerExtendedCommonCheckers() {
     useCheckers(ExtendedExpressionCheckers)
     useCheckers(ExtendedDeclarationCheckers)
 }
+
+// TODO: Move this to different, JVM-specific module?
+fun FirSessionFactory.FirSessionConfigurator.registerJvmCheckers() {
+    useCheckers(JvmDeclarationCheckers)
+}
